make MdqVariantsChecker a generic VariantsChecker
