+++ src/tree_test_utils.rs
@@ -4,13 +4,6 @@ pub use test_utils::*;
 // export its contents.
 #[cfg(test)]
 mod test_utils {
-    use std::collections::HashSet;
-    use std::sync::{Arc, Mutex};
-    use std::{thread, time};
-
-    use regex::Regex;
-
-    use crate::tree::MdqNode;
 
     macro_rules! nodes_matcher {
         {$($variant:ident$(($payload:pat))?),* $(,)?} => {
@@ -66,63 +59,5 @@ mod test_utils {
                 value: $text.to_string(),
             }
         };
-    }
-
-    pub struct MdqVariantsChecker {
-        require: Arc<Mutex<HashSet<String>>>,
-    }
-
-    // TODO unify this with the one in tree.rs.
-    // I can do that by making this generic, as well as making nodes_matcher! work with generics by having it pass in
-    // the type as an arg (e.g. `mdast::Node` or `MdqNode`
-    impl MdqVariantsChecker {
-        pub fn new() -> Self {
-            let all_node_names = nodes_matcher! {
-                Root(_),
-                Header(_),
-                Paragraph(_),
-                BlockQuote(_),
-                List(_),
-                Table(_),
-                ThematicBreak,
-                CodeBlock(_),
-                Inline(_),
-            };
-            Self {
-                require: Arc::new(Mutex::new(all_node_names)),
-            }
-        }
-
-        pub fn see(&self, node: &MdqNode) {
-            let node_debug = format!("{:?}", node);
-            let re = Regex::new(r"^\w+").unwrap();
-            let node_name = re.find(&node_debug).unwrap().as_str();
-            self.require.lock().map(|mut set| set.remove(node_name)).unwrap();
-        }
-
-        pub fn wait_for_all(&self) {
-            let timeout = time::Duration::from_millis(500);
-            let retry_delay = time::Duration::from_millis(50);
-            let start = time::Instant::now();
-            loop {
-                if self.require.lock().map(|set| set.is_empty()).unwrap() {
-                    break;
-                }
-                if start.elapsed() >= timeout {
-                    let mut remaining: Vec<String> = self
-                        .require
-                        .lock()
-                        .map(|set| set.iter().map(|s| s.to_owned()).collect())
-                        .unwrap();
-                    remaining.sort();
-                    panic!(
-                        "Timed out, and missing {} variants:\n- {}",
-                        remaining.len(),
-                        remaining.join("\n- ")
-                    )
-                }
-                thread::sleep(retry_delay);
-            }
-        }
     }
 }
